pyproject.toml
[tool.poetry]
name = "sqlark"
version = "0.0.0" # placeholder version (versioning handled by poetry dynamic versioning)
description = "A lightweight Non-ORM with a builder pattern for querying a postgres database"
authors = ["Joshua Shapiro <jshapiro@meadowlarkengineering.com>"]
readme = "README.md"
packages = [{include = "sqlark", from = "src"}]

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
bump = true

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
boto3 = "^1.35.29"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pylint = "^3.0.4"
pytest-postgresql = "^4.1.1"
pytest-mock = "^3.11.1"
pre-commit = "^4.1.0"


[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.pylint]
max-line-length = 120
disable = [
    "R0913", # (too-many-arguments)
    "R0917", # (too-many-positional-arguments)
    "R0903", # (too-few-public-methods)
    "C0103", # (invalid-name)
    "W0622", # (redefined-builtin)
    "C3001", # (unnecessary-lambda-assignment)
    "W0105", # (pointless-string-statement)
    "R0401", # (cyclic-import)
    "R0801", # (similarities)
]
